Chatpro/requirements.txt
asgiref==3.7.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==3.13.0
django-location-field==2.1.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
idna==3.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
phonenumberslite==8.13.0
platformdirs==2.5.3
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
virtualenv==20.16.6
